Revert "fixup"
This reverts commit d9ff9f58fe7e2359079dc3c34b3ef8442bfc254a.

+++ tools/gluten-te/centos/centos-7-deps.sh
@@ -48,15 +48,14 @@ pip3 install --upgrade pip
 pip3 install cmake==3.28.3
 
 # git >= 2.7.4
-case "$(git --version)" in "git version 2."*)
+if [[ "$(git --version)" != "git version 2."* ]]; then
   [ -f /etc/yum.repos.d/ius.repo ] || yum -y install https://repo.ius.io/ius-release-el7.rpm
   yum -y remove git
   yum -y install git236
-  ;;
-esac
+fi
 
 # flex>=2.6.0
-case "$(PATH="/usr/local/bin:$PATH" flex --version 2>&1)" in "flex 2.6."*)
+if [[ "$(PATH="/usr/local/bin:$PATH" flex --version 2>&1)" != "flex 2.6."* ]]; then
   yum -y install gettext-devel
   FLEX_URL="https://github.com/westes/flex/releases/download/v2.6.4/flex-2.6.4.tar.gz"
   mkdir -p /tmp/flex
@@ -67,8 +66,7 @@ case "$(PATH="/usr/local/bin:$PATH" flex --version 2>&1)" in "flex 2.6."*)
   make install
   cd
   rm -rf /tmp/flex
-  ;;
-esac
+fi
 
 # automake>=1.14
 installed_automake_version="$(aclocal --version | sed -En "1s/^.* ([1-9\.]*)$/\1/p")"
